Reverting back. Clock time is long one more time, and time delta is int

+++ engine/src/main/kotlin/star/galaxy/engine/Clock.kt
@@ -4,8 +4,8 @@ import star.galaxy.engine.metainformations.Second
 
 interface Clock {
     @Second
-    fun next(): Double
+    fun next(): Int
 
     @Second
-    fun time(): Double
+    fun time(): Long
 }
