+++ engine/src/main/kotlin/star/galaxy/engine/physics/ForceApplicator.kt
@@ -5,7 +5,7 @@ import star.galaxy.engine.types.ForceApplicable
 import star.galaxy.engine.types.ForceApplicables
 
 interface ForceApplicator {
-    fun apply(obj: ForceApplicables, @Second Δt: Double)
+    fun apply(obj: ForceApplicables, @Second Δt: Int)
 
-    fun apply(obj: ForceApplicable, @Second Δt: Double)
+    fun apply(obj: ForceApplicable, @Second Δt: Int)
 }
